add: "Critical" options
Even after previous attempts to fix, I was still encountering situations when `TracerLogFile` would not correctly close a json array on script exit. I realized this was being caused by the thread being interrupted by the message queue. Including a call to `Critical` fixed the issue, so I exposed the values as options to provide the user with control over what gets passed to `Critical`